# PERF: PeriodEngine.get_loc accept narrow data type