@pep8speaks 2019-10-26
Hello @proost! Thanks for opening this PR. We checked the lines you've touched for [PEP 8](https://www.python.org/dev/peps/pep-0008) issues, and found:

* In the file [`pandas/tests/indexes/period/test_indexing.py`](https://github.com/pandas-dev/pandas/blob/ac81f60a5c4c1943f3eb855f8fc288ae772dcf64/pandas/tests/indexes/period/test_indexing.py):

> [Line 694:51](https://github.com/pandas-dev/pandas/blob/ac81f60a5c4c1943f3eb855f8fc288ae772dcf64/pandas/tests/indexes/period/test_indexing.py#L694): [W605](https://duckduckgo.com/?q=pep8%20W605) invalid escape sequence '\('
> [Line 694:68](https://github.com/pandas-dev/pandas/blob/ac81f60a5c4c1943f3eb855f8fc288ae772dcf64/pandas/tests/indexes/period/test_indexing.py#L694): [W605](https://duckduckgo.com/?q=pep8%20W605) invalid escape sequence '\)'