@jbrockmendel 2019-10-27
Is there any simplficiation of PeriodIndex.get_loc that this makes possible?