@proost 2019-10-30
@jbrockmendel 

In terms of simplication, Yes. I was wrong.
How about memory Usage?
Before,
```
In [2]: pi = pd.period_range(start=2010,end=2019,freq="m")                      
In [3]: pi.memory_usage()                                                       
Out[3]: 872
In [4]: pi.get_loc(pi[0])                                                       
Out[4]: 0
In [5]: pi.memory_usage()                                                       
Out[5]: 11984
```
After
```
In [3]: pi = pd.period_range(start=2010,end=2019,freq="m")                                                                                      
In [4]: pi.memory_usage()                                                                                                                       
Out[4]: 872
In [5]: pi.get_loc(pi[0])                           
Out[5]: 0
In [6]: pi.memory_usage()                           
Out[6]: 6864
```

I don't intent to it but i found it running test case. After `.get_loc`, use less memory space in each column.