@jbrockmendel 2019-10-31
@proost I think you need to clarify the goal of this PR.  The referenced issue #29038 is about the behavior of PeriodEngine, but this doesn't touch that at all.