@proost 2019-10-31
@jbrockmendel 
Purpose: Improvement function of `PeriodIndex.get_loc` 
Reason: Inefficient `PeriodIndex.get_loc`. Because `PeriodIndex.get_loc` acceptable parameter type is narrow.  So most case `PeriodIndex.get_loc` compute twice. Both `PeriodEngine` and `Int64Index`
So What i try to : 
-> first : Change `PeriodEngine.get_loc` to accept `Period` to cover all parameter types. Problem is `PeriodEngine.get_loc` can't accept "tolerance" and "method" 
-> second: Don't use `PeriodEngine.get_loc` Using `PeriodIndex,get_loc` .