@jbrockmendel 2019-11-19
@proost are there non-trivial performance improvements from this?  If not, you need to clarify what this is accomplishing.