@proost 2019-11-20
@jbrockmendel 
small improvement is prohibited? If prohibited, I will close this PR. If not, this PR for reducing memory usage.