@jbrockmendel 2019-11-21
Small improvements are encouraged!  We just need to communicate clearly what the relevant improvement is.  how much less memory does this use?  You're no longer calling the _engine method at the beginning, which makes me concerned that performance will take a hit.