@proost 2019-11-21
@jbrockmendel 
After using get_loc, period index use less memory space before about 60% ~ 100%

first case: 
'''
pi = pd.period_range(start=2000,end=2019,freq="y") 
pi.get_loc('2000')
pi.memory_usage()
'''
Before change, period index memory usage: 1600
After change, period index memory usage: 900

second case:
'''
pi = pd.period_range(start=2000,end=2019,freq="m") 
pi.get_loc('2000')
pi.memory_usage()
'''
Before change, period index memory usage: 24144
After change, period index memory usage: 13904

third case:
'''
pi = pd.period_range(start=2010,end=2019,freq="d") 
pi.get_loc('2010')
pi.memory_usage()
'''
Before change, period index memory usage: 380288
After change, period index memory usage: 216448

fourth case:
'''
pi = pd.period_range(start=2010,end=2019,freq="h") 
pi.get_loc('2010')
pi.memory_usage()
'''
Before change, period index memory usage: 6505104
After change, period index memory usage: 3883664

fifth case:
'''
pi = pd.period_range(start=2010,end=2019,freq="s") 
pi.get_loc('2010')
pi.memory_usage()
'''
Before change, period index memory usage: 4543948816
After change, period index memory usage: 4543948816

if you worry about time performance from this change, no need to worry.
command same above, 

first case:
'''
Before Change: 11.7 µs ± 58.3 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)
After Change: 10.2 µs ± 42.6 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)
'''

second case:
'''
Before Change: 11.7 µs ± 2.17 µs per loop (mean ± std. dev. of 7 runs, 100000 loops each)
After Change: 8.78 µs ± 31.5 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)
'''

third case:
'''
Before Change: 11 µs ± 384 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)
After Change: 9.35 µs ± 40.5 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)
'''

fourth case:
'''
Before Change: 10.7 µs ± 330 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)
After Change: 9.25 µs ± 216 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)
'''

fifth case:
'''
Before Change: 28.8 µs ± 17.5 µs per loop (mean ± std. dev. of 7 runs, 1 loop each)
After Change: 26.8 µs ± 17.4 µs per loop (mean ± std. dev. of 7 runs, 1 loop each)
'''

Almost same or slightly better before. So this change of code doesn't exacerbate on time performance.