@jreback 2019-12-02
so i wonder after this if we can blow PeriodEngine away entirely. cc @jbrockmendel 

@proost the memory savings are because we no longer need to instantiate the engine.