@jbrockmendel 2019-12-04
@proost if i understand correclty, all the timings you posted were with string inputs.  Can you do the same comparisons for 0, pd.Period("2000"), pd.NaT, pd.Timedelta(days=1)