@jbrockmendel 2019-12-04
> so i wonder after this if we can blow PeriodEngine away entirely

It looks like there is some non-trivial stuff in there for `get_indexer`, but we might be able to share more with DatetimeEngine