@proost 2019-12-09
@jbrockmendel 
If
```
l = ['1970-01-01', '1970-01-02', '1970-01-03', '1970-01-01',
     '1970-01-05', '1970-01-06', '1970-01-07', '1970-01-08',
     '1970-01-09', '1970-01-10', '1970-01-11', '1970-01-12',
     '1970-01-13', '1970-01-14', '1970-01-15', '1970-01-16',
     '1970-01-17', '1970-01-18', '1970-01-19', '1970-01-20',
     '1970-01-21', '1970-01-22', '1970-01-23', '1970-01-24',
     '1970-01-25', '1970-01-26', '1970-01-27', '1970-01-28',
     '1970-01-29', '1970-01-30', '1970-01-31', np.datetime64('NaT')
]
pi = pd.PeriodIndex(l,freq="D")
```

case 1: 
```
p = pd.Period(year=1970, month=1, day=1, freq="D")
pi.get_loc(p)
```
Before:
memory usage: 3072
time taken: 11.8 µs ± 128 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)
After:
memory usage: 1792
time taken: 14 µs ± 101 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)

case 2:
```
pi.get_loc(0)
```
Before: 
memory usage: 1536
time taken:  5.13 µs ± 25 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)
After:
memory usage: 1792
time taken: 11.2 µs ± 2.36 µs per loop (mean ± std. dev. of 7 runs, 100000 loops each)

case 3:
```
pi.get_loc(pd.NaT)
```
Before:
memory usage: 3072
time taken: 10.1 µs ± 348 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each) 
After:
memory usage: 1792
time taken: 11.7 µs ± 51.4 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)

case 4:
Because "Timedelta" Object can't be converted to Period. So can't work on ".get_loc" 
But if you mean "Timestamp" object,
```
t = pd.Timestamp(year=1970, month=1, day=1)
pi.get_loc(t)
```
Before:
memory usage: 3072
time taken: 13.3 µs ± 533 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)
After:
memory Usage: 1792
time taken: 16 µs ± 2.59 µs per loop (mean ± std. dev. of 7 runs, 100000 loops each)

Well, time performance becomes bad before. But i think this much can be acceptable.
except Integer case, all memory usage lower before. And Integer or float type is not expected input from user. Becasue `PeriodIndex.get_loc` parameter type is expected as "Period", "String", "datetime" or datetimelike object from user. So This is not a problem.